backend/requirements.txt
accelerate==1.9.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
bitsandbytes==0.45.0
nh3==0.2.15
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
distro==1.9.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
Flask-Cors==3.0.10
fsspec==2025.7.0
gevent==24.11.1
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.0
idna==3.10
inflect==7.3.1
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.93.0
packaging==25.0
parso==0.8.4
peft==0.15.0
pexpect==4.9.0
pillow==11.3.0
pip==25.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyClamd==0.4.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.21.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.1.0
typeguard==4.4.4
typing_extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.1
google-generativeai==0.8.3
argostranslate==1.9.0
Flask-Babel==4.0.0
